Renamed package name and components to match current terms.
