+++ demo-app/app/src/main/java/com/launchkey/android/authenticator/demo/ui/BaseDemoView.java
@@ -1,0 +1,11 @@
+package com.launchkey.android.authenticator.demo.ui;
+
+
+import com.launchkey.android.authenticator.sdk.AuthenticatorManager;
+
+/**
+ * Created by armando on 7/8/16.
+ */
+public interface BaseDemoView {
+    AuthenticatorManager getAuthenticatorManager();
+}
