--- demo-app/app/src/main/java/com/launchkey/android/whitelabel/demo/ui/fragment/CustomUnlinkFragment2.java
@@ -1,55 +1,0 @@
-package com.launchkey.android.whitelabel.demo.ui.fragment;
-
-import android.app.ProgressDialog;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-
-import com.launchkey.android.authenticator.sdk.AuthenticatorManager;
-import com.launchkey.android.authenticator.sdk.SimpleOperationCallback;
-import com.launchkey.android.authenticator.sdk.error.BaseError;
-import com.launchkey.android.whitelabel.demo.R;
-import com.launchkey.android.whitelabel.demo.util.Utils;
-
-/**
- * Created by armando on 7/20/16.
- */
-public class CustomUnlinkFragment2 extends BaseDemoFragment {
-
-    private ProgressDialog mUnlinkingDialog;
-
-    @Override
-    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-
-        mUnlinkingDialog = new ProgressDialog(getActivity(), R.style.Theme_WhiteLabel_Dialog);
-        mUnlinkingDialog.setIndeterminate(true);
-        mUnlinkingDialog.setCancelable(false);
-        mUnlinkingDialog.setMessage("Custom Unlinking...");
-
-        unlink();
-    }
-
-    private void unlink() {
-
-        Utils.show(mUnlinkingDialog);
-
-        AuthenticatorManager.getInstance().unlinkCurrentDevice(new SimpleOperationCallback() {
-
-            @Override
-            public void onResult(boolean successful, BaseError error, Object o) {
-
-                Utils.dismiss(mUnlinkingDialog);
-
-                if (successful) {
-                    if (getActivity() != null && !getActivity().isFinishing()) {
-                        getActivity().finish();
-                    }
-                } else {
-
-                    String message = Utils.getMessageForBaseError(error);
-                    Utils.alert(getActivity(), "Error", message);
-                }
-            }
-        });
-    }
-}
