+++ demo-app/app/src/main/res/values/strings.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">White Label Demo</string>
+    <string name="app_name">Auth Demo</string>
 
     <string name="demo_generic_ok">OK</string>
 
-    <string name="demo_activity_list_title_format">WL Demo (%1$s)</string>
+    <string name="demo_activity_list_title_format">Auth Demo (%1$s)</string>
     <string name="demo_activity_authrequest_title">Auth Request</string>
     <string name="demo_activity_otps_title">One-Time Passwords</string>
